添加docker2不再维护的说明，迁移过期脚本(狂欢城 jd_818.js),iOS订阅也进行更新
